Improved CSEG and SSEG; fixed minor exception in CSEG.

+++ SharedUtils/SharedUtils.hpp
@@ -155,6 +155,6 @@ std::string get_typeid_name_of_class(const T& t) {
 }
 
 std::string ConvertOrientationToString(double orientation);
-std::string ConvertColorToString(double r, double g, double b);
+std::string ConvertColorToString(float r, float g, float b);
 
 #endif
